- fixed warning @FlexRayScheduler.cpp
git-archimport-id: falk@cs.fau.de--2008-systemoc-intern/SystemC-VPC--multihop--0.6--patch-99

+++ FlexRayScheduler.cpp
@@ -349,7 +349,7 @@ namespace SystemC_VPC{
     }
     
     if(processcount==0 /*&& firstrun == false */){
-	    cyclecount = sc_time_stamp() / cycle_length;
+	    cyclecount = (int) (sc_time_stamp() / cycle_length);
 	    to_init=true;
 	   // cout<<"new cyclecount: "<<cyclecount<<endl;
      }
